vims updates

+++ backend/hexalyte-vims-api-v1-10/migrations/20250101000011-create-productstorage.js
@@ -22,7 +22,7 @@ module.exports = {
         }
       },
       Quantity: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL(10, 2)
       },
       LastUpdated: {
         type: Sequelize.DATE
