+++ backend/hexalyte-vims-api-v1-10/models/productstorage.js
@@ -32,7 +32,7 @@ module.exports = (sequelize, DataTypes) => {
         key: 'LocationID'
       }
     },
-    Quantity: DataTypes.INTEGER,
+    Quantity: DataTypes.DECIMAL(10, 2),
     LastUpdated: DataTypes.DATE
   }, {
     sequelize,
